intial next js setup
